backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.60.0
sentence-transformers==3.2.0
numpy==2.1.3
scikit-learn==1.5.2
python-pptx==1.0.2
python-multipart==0.0.20
pyjwt==2.8.0
